Speed up unit tests and reduce duplicated code
