+++ application/change_event_dlq_handler/tests/test_change_event_dlq_handler.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass
 from json import dumps
 from typing import Any
 from unittest.mock import MagicMock, patch
@@ -126,20 +125,6 @@ def dead_letter_change_event_from_holding_queue():
             },
         ],
     }
-
-
-@pytest.fixture()
-def lambda_context():
-    @dataclass
-    class LambdaContext:
-        """Mock LambdaContext - All dummy values."""
-
-        function_name: str = "change-event-dlq-handler"
-        memory_limit_in_mb: int = 128
-        invoked_function_arn: str = "arn:aws:lambda:eu-west-1:000000000:function:change-event-dlq-handler"
-        aws_request_id: str = "52fdfc07-2182-154f-163f-5f0f9a621d72"
-
-    return LambdaContext()
 
 
 @patch(f"{FILE_PATH}.extract_body")
