+++ application/service_sync/tests/test_service_sync.py
@@ -1,9 +1,7 @@
-from dataclasses import dataclass
 from json import dumps
 from os import environ
 from unittest.mock import MagicMock, call, patch
 
-import pytest
 from aws_lambda_powertools.logging import Logger
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
@@ -41,20 +39,6 @@ SQS_EVENT = {
         },
     ],
 }
-
-
-@pytest.fixture()
-def lambda_context():
-    @dataclass
-    class LambdaContext:
-        """Mock LambdaContext - All dummy values."""
-
-        function_name: str = "service-sync"
-        memory_limit_in_mb: int = 128
-        invoked_function_arn: str = "arn:aws:lambda:eu-west-1:000000000:function:service-sync"
-        aws_request_id: str = "52fdfc07-2182-154f-163f-5f0f9a621d72"
-
-    return LambdaContext()
 
 
 @patch(f"{FILE_PATH}.check_and_remove_pending_dos_changes")
